LayoutTests/inspector/protocol/console-agent-expected.txt

CONSOLE MESSAGE: line 1: test
CONSOLE MESSAGE: line 1: test
Protocol stability test. It is dumping request/response pairs of ConsoleAgent functions.

-----------------------------------------------------------
ConsoleAgent.disable()

request:
{
    method : "Console.disable"
    id : <number>
}

response:
{
    result : {
    }
    id : <number>
}

-----------------------------------------------------------
ConsoleAgent.enable()

request:
{
    method : "Console.enable"
    id : <number>
}

response:
{
    result : {
    }
    id : <number>
}

-----------------------------------------------------------
RuntimeAgent.evaluate("console.info('test'); console.info('test')","test-group",false)

request:
{
    method : "Runtime.evaluate"
    params : {
        expression : "console.info('test'); console.info('test')"
        objectGroup : "test-group"
        includeCommandLineAPI : false
    }
    id : <number>
}

event ConsoleAgent.messageAdded
{
    method : "Console.messageAdded"
    params : {
        message : {
            source : "console-api"
            level : "log"
            text : "test"
            type : "log"
            line : 1
            url : ""
            repeatCount : 1
            parameters : [
                {
                    type : "string"
                    value : "test"
                }
            ]
            stackTrace : <object>
        }
    }
}

event ConsoleAgent.messageRepeatCountUpdated
{
    method : "Console.messageRepeatCountUpdated"
    params : {
        count : 2
    }
}

response:
{
    result : {
        result : {
            type : "undefined"
        }
    }
    id : <number>
}

-----------------------------------------------------------
ConsoleAgent.clearMessages()

request:
{
    method : "Console.clearMessages"
    id : <number>
}

event ConsoleAgent.messagesCleared
{
    method : "Console.messagesCleared"
}

response:
{
    result : {
    }
    id : <number>
}

-----------------------------------------------------------
ConsoleAgent.setMonitoringXHREnabled(true)

request:
{
    method : "Console.setMonitoringXHREnabled"
    params : {
        enabled : true
    }
    id : <number>
}

response:
{
    result : {
    }
    id : <number>
}

-----------------------------------------------------------
ConsoleAgent.setMonitoringXHREnabled(false)

request:
{
    method : "Console.setMonitoringXHREnabled"
    params : {
        enabled : false
    }
    id : <number>
}

response:
{
    result : {
    }
    id : <number>
}

-----------------------------------------------------------
ConsoleAgent.addInspectedNode(<number>)

request:
{
    method : "Console.addInspectedNode"
    params : {
        nodeId : 1
    }
    id : <number>
}

response:
{
    result : {
    }
    id : <number>
}

===========================================================
Coverage for ConsoleAgent
{
    enable : "checked"
    disable : "checked"
    clearMessages : "checked"
    setMonitoringXHREnabled : "checked"
    addInspectedNode : "checked"
    messageAdded : "checked"
    messageRepeatCountUpdated : "checked"
    messagesCleared : "checked"
}
